fix: 댓글 컴포넌트 연결, 쿼리 기능 타입오류 수정
