+++ src/app/queries/post/getPostById.ts
@@ -1,7 +1,19 @@
 'use server';
 
-import { PostWithSchedule } from '@hooks/post/useFetchPost';
+import { Database } from '@ts/supabase';
 import { createClient } from '@utils/supabase/server';
+
+type PostType = Pick<Database['public']['Tables']['posts']['Row'], 'id' | 'content'>;
+type ScheduleType = Pick<
+  Database['public']['Tables']['schedules']['Row'],
+  'name' | 'memo' | 'created_at' | 'start_date' | 'end_date' | 'start_time'
+>;
+
+export type PostWithSchedule = {
+  id: PostType['id'];
+  content: PostType['content'];
+  schedules: ScheduleType[];
+};
 
 // 게시글 불러오기
 export const getPost = async (postId: string): Promise<PostWithSchedule> => {
